requirements.txt
torch
numpy
librosa
openai-whisper
pyannote.audio
python-dotenv
Flask